pyproject.toml
[project]
name = "github-mcp-shim"
version = "0.1.0"
description = "A stdio-to-http shim for using GitHub Apps with the GitHub MCP server"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
dependencies = [
    "httpx>=0.27.0",
    "pyjwt>=2.8.0",
    "cryptography>=42.0.0",
]

[project.scripts]
github-mcp-shim = "github_mcp_shim.__main__:main_entry"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
